app: add initial simulator configuration.
This includes the set of plugins currently compatible with GUIs
available at LNLS Sirius beamlines. The exception here is ffmpegServer
and ffmpegFile plugins, which are outside ADCore and have not been
added to EPICS in Docker yet.

`CONFIG_SITE` from `areaDetector` support module has been included in
the configuration to make use of the support libraries definition used
in the commonDriverMakefile definitions.