boot: add iocs script configuration in start-up example.
The `st.cmd` iocsh script can be defined to support the SIRIUS iocs
script without affecting the users that don't use it. Since it is
convenient and commonly used in the SIRIUS beamlines, we now add support
for it.

Changing directory is used instead of absolute paths to prevent updating
all occurrences of relative file definitions while having the same
effect. If local files are to be used (from the default mounted volume),
they should use absolute paths instead.